reorganize files in compiler dir
